fix: handle duplicate error instead of sql error fix: set fqdns to null if you delete an app or a serviceapp fix: make stucked resources a separate command

+++ app/Actions/Application/StopApplication.php
@@ -11,6 +11,9 @@ class StopApplication
     public function handle(Application $application)
     {
         $server = $application->destination->server;
+        if (!$server->isFunctional()) {
+            return 'Server is not functional';
+        }
         if ($server->isSwarm()) {
             instant_remote_process(["docker stack rm {$application->uuid}" ], $server);
         } else {
