+++ app/Actions/Database/StopDatabase.php
@@ -17,6 +17,9 @@ class StopDatabase
     public function handle(StandaloneRedis|StandalonePostgresql|StandaloneMongodb|StandaloneMysql|StandaloneMariadb $database)
     {
         $server = $database->destination->server;
+        if (!$server->isFunctional()) {
+            return 'Server is not functional';
+        }
         instant_remote_process(
             ["docker rm -f {$database->uuid}"],
             $server
